Allow passing popover config via data attribute
