+++ packages/popover/src/js/defaults.js
@@ -1,6 +1,9 @@
 export default {
   autoMount: false,
 
+  // Data attributes
+  dataConfig: "popover-config",
+
   // Selectors
   selectorPopover: ".popover",
   selectorTooltip: ".popover_tooltip",
