+++ packages/popover/src/js/open.js
@@ -1,4 +1,4 @@
-import { getConfig, getModifiers, getPopover } from "./helpers";
+import { getCustomPropConfig, getModifiers, getPopover } from "./helpers";
 
 export async function open(query) {
   // Get the popover from collection.
@@ -14,7 +14,7 @@ export async function open(query) {
   }
 
   // Update popover config.
-  popover.config = getConfig(popover.el, this.settings);
+  popover.config = getCustomPropConfig(popover.el, this.settings);
 
   // Enable popper event listeners and set placement/modifiers.
   popover.popper.setOptions({
